# docs/swagger.yaml
basePath: /
definitions:
...
  templates.CreateTemplateRequest:
    properties:
      category:
        type: string
      design_json:
        additionalProperties: true
        type: object
      name:
        type: string
      thumbnail:
        type: string
    required:
    - category
    - design_json
    - name
    type: object
  templates.Template:
    properties:
      category:
        type: string
      created_at:
        type: string
      design_json:
        additionalProperties: true
        description: frame, colors, shapes, logo position
        type: object
      id:
        type: string
      name:
        type: string
      thumbnail:
        description: image URL
        type: string
      user_id:
        description: null = global template
        type: string
    type: object
  templates.UpdateTemplateRequest:
    properties:
      category:
        type: string
      design_json:
        additionalProperties: true
        type: object
      name:
        type: string
      thumbnail:
        type: string
    type: object
  user.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Scalable QR Code SaaS backend with dynamic QR, analytics, billing,
    templates, folders, OAuth, and more.
  title: QR SaaS API
  version: "2.0"
paths:
...
  /api/templates/:
    post:
      consumes:
      - application/json
      parameters:
      - description: template
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/templates.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/templates.Template'
      security:
      - BearerAuth: []
      summary: Create a template
      tags:
      - Templates
  /api/templates/{id}:
    delete:
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete template
      tags:
      - Templates
    get:
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/templates.Template'
      security:
      - BearerAuth: []
      summary: Get template
      tags:
      - Templates
    put:
      consumes:
      - application/json
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: update
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/templates.UpdateTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/templates.Template'
      security:
      - BearerAuth: []
      summary: Update template
      tags:
      - Templates
...
  /api/templates/mine:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/templates.Template'
            type: array
      security:
      - BearerAuth: []
      summary: List user templates
      tags:
      - Templates
  /api/user/me:
    get:
      description: Returns user profile for current JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get logged-in user
      tags:
      - User
...
swagger: "2.0"